requirements.txt
# Optimized requirements with performance-focused libraries
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Compress==1.13  # Response compression
Flask-Limiter==3.5.0  # Rate limiting

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7  # PostgreSQL adapter (optional)

# Caching
redis==4.6.0
flask-caching==2.1.0

# Performance monitoring
psutil==5.9.5  # System monitoring
gunicorn==21.2.0  # Production WSGI server

# HTTP client (for external API calls)
httpx==0.24.1  # Async HTTP client
requests==2.31.0

# HTML parsing (for web scraping)
beautifulsoup4==4.12.2

# Data processing (temporarily disabled due to numpy compatibility)
pandas==1.5.3
openpyxl==3.1.2  # Excel 파일 읽기용
lxml==4.9.3

# Background tasks
APScheduler==3.10.4
celery==5.2.7  # Optional for heavy background tasks (Python 3.9 compatible)
schedule==1.2.0  # Simple job scheduling

# JSON processing
orjson==3.9.7  # Faster JSON parsing

# Security
cryptography==41.0.4
PyJWT==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development (optional)
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Memory profiling (optional)
memory-profiler==0.61.0
pympler==0.9

# Environment management
python-dotenv==1.0.0
